更改detail文件夹为core文件夹
