+++ include/aquarius/core/deadline_timer.hpp
@@ -4,7 +4,7 @@
 
 namespace aquarius
 {
-	namespace detail
+	namespace core
 	{
 		using deadline_timer =  boost::asio::basic_deadline_timer<std::chrono::system_clock::time_point,time_traits<std::chrono::system_clock::time_point>>;
 	}
