+++ include/aquarius/core/router.hpp
@@ -6,7 +6,7 @@
 
 namespace aquarius
 {
-	namespace detail
+	namespace core
 	{
 		template<typename R, typename... Args>
 		class router
@@ -47,7 +47,7 @@ namespace aquarius
 		template<typename R, typename... Args>
 		class single_router
 			: public router<R, Args...>
-			, public detail::singleton<router<R, Args...>>
+			, public core::singleton<router<R, Args...>>
 		{
 
 		};
